Split TypeScript abort tests
